build / deploy update:
- converted deploy process to v2.0
- due to no heroku integration yet had to follow the docs here:
https://circleci.com/docs/2.0/deployment_integrations/#heroku
- minor ci fix
- updated run-tests for local use which properly sets up and cleans up

+++ run-tests-local.sh
@@ -1,5 +1,6 @@
 #!/bin/sh
 
-dropdb listopia-test
+[[ -n $(psql -l | grep listopia-test) ]] && dropdb listopia-test
 createdb listopia-test
 lein with-profile test migratus migrate && lein test
+dropdb listopia-test
